+++ flecsi/execution.hh
@@ -394,9 +394,11 @@ reduce(Args &&... args) {
   std::size_t & flog_task_count = flecsi_context.flog_task_count();
   ++flog_task_count;
   if(flog_task_count % FLOG_SERIALIZATION_INTERVAL == 0 &&
-     reduce_internal<log::log_size, fold::max, flecsi::mpi>().get() >
-       FLOG_SERIALIZATION_THRESHOLD)
-    reduce_internal<log::send_to_one, void, flecsi::mpi>();
+     reduce_internal<log::state::log_size, fold::max, flecsi::mpi>(
+       *log::state::instance)
+         .get() > FLOG_SERIALIZATION_THRESHOLD)
+    reduce_internal<log::state::gather, void, flecsi::mpi>(
+      *log::state::instance);
 #endif
 
   return reduce_internal<Task, Reduction, Attributes, Args...>(
@@ -450,7 +452,8 @@ namespace log {
 inline void
 flush() {
 #if defined(FLECSI_ENABLE_FLOG) && defined(FLOG_ENABLE_MPI)
-  flecsi::exec::reduce_internal<log::send_to_one, void, flecsi::mpi>();
+  flecsi::exec::reduce_internal<log::state::gather, void, flecsi::mpi>(
+    *log::state::instance);
   flecsi::run::context::instance().flog_task_count() = 0;
 #endif
 } // flush