+++ flecsi/log/utils.hh
@@ -124,18 +124,6 @@ rstrip(const char * file) {
   return tmp.substr(tmp.rfind(C) + 1);
 } // rstrip
 
-/*!
-  Returnt the current number of buffered packets.
- */
-
-std::size_t log_size();
-
-/*!
-  Serialize output to a single process.
- */
-
-void send_to_one();
-
 /// \}
 } // namespace log
 } // namespace flecsi